+++ core/src/main/java/de/bitbrain/braingdx/ui/Navigateable.java
@@ -1,0 +1,7 @@
+package de.bitbrain.braingdx.ui;
+
+public interface Navigateable {
+   void previous();
+   void next();
+   void enter();
+}